pyproject.toml
[project]
name = "weight-of-reasoning"
version = "0.1.0"
description = "Reproducible research codebase to compute internal reasoning effort metrics for open LLMs"
authors = [
    {name = "Research Team", email = "research@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch==2.2.2",
    "transformers>=4.30.0",
    "datasets>=2.12.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "transformer-lens>=2.0.0,<3.0.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "pingouin>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wor"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
